+++ src/services/query_service/app/dtos/valuation_dto.py
@@ -6,15 +6,15 @@ class ValuationData(BaseModel):
     """
     Represents the valuation details for a position.
     """
-    market_price: Decimal | None
+    market_price: float | None
     
     # In portfolio base currency
-    market_value: Decimal | None
-    unrealized_gain_loss: Decimal | None
+    market_value: float | None
+    unrealized_gain_loss: float | None
     
     # In instrument's local currency
-    market_value_local: Decimal | None
-    unrealized_gain_loss_local: Decimal | None
+    market_value_local: float | None
+    unrealized_gain_loss_local: float | None
 
     model_config = ConfigDict(
         from_attributes=True
